+++ server/src/main/java/org/opensearch/cluster/service/ClusterService.java
@@ -292,6 +292,17 @@ public class ClusterService extends AbstractLifecycleComponent {
     }
 
     /**
+     * Functionality for register task key to cluster manager node.
+     *
+     * @param taskKey - task key of task
+     * @param throttlingEnabled - throttling is enabled for task or not i.e does data node perform retries on it or not
+     * @return throttling task key which needs to be passed while submitting task to cluster manager
+     */
+    public ClusterManagerTaskThrottler.ThrottlingKey registerClusterManagerTask(String taskKey, boolean throttlingEnabled) {
+        return clusterManagerService.registerClusterManagerTask(taskKey, throttlingEnabled);
+    }
+
+    /**
      * Submits a cluster state update task; unlike {@link #submitStateUpdateTask(String, Object, ClusterStateTaskConfig,
      * ClusterStateTaskExecutor, ClusterStateTaskListener)}, submitted updates will not be batched.
      *